fix: Correctly parse GTS if it contains NaN
NaN is not recognize as a json object which is turned into a NULL object.
But to transform a gts into a dataframe, if values have some NULL values, then the end result would be that the value column would have a list of list instead of list of double/long/...
To take that into account, transform NULL to NA before the transformation.